requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
colorlog==4.4.0
colormap==1.0.3
configparser==5.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
CurrencyConverter==0.14.3
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-flower==1.0.0
django-jinja==2.7.0
django-jsonfield==1.4.0
django-kronos==1.0
django-phonenumber-field==5.0.0
django-picklefield==3.0.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-resetpassword==0.1.2
django-sendsms==0.4
django-smsish2==1.3.5
django-templated-mail==1.1.1
django-uuidfield-2==0.6.6
djangorestframework==3.12.1
djoser==2.1.0
easydev==0.10.1
flower==0.9.5
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
httplib2==0.18.1
humanize==0.5.1
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
pexpect==4.8.0
phonenumbers==8.12.11
Pillow==8.0.1
plumbum==1.6.9
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
PyGoogleVoice==0.5.2
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-rapidjson==0.9.4
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rpyc==4.1.5
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tornado==6.0.4
twilio==6.45.4
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
zope.event==4.5.0
zope.interface==5.1.2
